Convert the forms of E4 mod. assignments back to E4(V, V, C, ...) forms
We have the parser returning a slightly simplified approach to E4 nodes, in that all modified assignment will always have the Operator based wrapping in it, no matter the type of the assignment. To help the generator, we don't need to put so much complexity into things like modified assignment, so we use a simpler form. This returns us to the simpler form from the parser.

Previously, this was done in the parser, before we unified assignment parsing with the rest of the schizo functions.