Fix an issue with double bindings and local resolution
We need to ensure that we are taking local resolutions ordered according to the appropriate evaluation order. If we don't, since we use only a single binding to anchor local bindings, it is possible that a local reference can be missed if an assignment to that local reference appears on the same line, but evaluated after the local reference, such as `i←i,i←∘∘∘`.