Reintroduce guard-first resolution
This is taken from the previous closure-based resolution. We didn't need this with the other design because of the way that bindings were calculated on dfns. However, since we have bindings in H nodes now, it's important that we account for their relative positions in the function body, and we have to also account for the guard paths, which may or may not introduce a separate timing for the localization of variables.

This means that we now have a single variable per `rf` value, rather than a single variable per `r`. This means that a singel dfns H node may have duplicate names, but that is because some may be scoped to the guards and some may be scoped to the main body. We may want to remove these duplicates later.

This approach still currently has the issue that it will not recognize a reference that is in the main body that precedes a guard assignment as the primary resolution node when resolving a reference inside of a guard statement, which may cause a name to miss a binding.