Make local resolution more intelligent
This adjusts our handling of local resolution to handle some particularly devious cases. Specifically, if we have a local binding that binds a name before entering a guard, but that guard also binds that name, but only after referencing said name, then the previous logic will default to finding that guard-local name, realize that it is not preceding the reference, and consider the reference free.

This new logic checks that a guard-local binding matches or does not, and if it does not, it will continue to try looking up to the local scope, instead of just giving up.

This adjustment should still preserve the semantics of our other lookup requirements, in particular, the requirement to be able to bind in a guard in a statement that occurs earlier in the text, followed by a free reference to the same name as the binding later.