Don't pre-emptively set the type of guard tests to array-values
We had originally set the type of guard tests to 1 pre-emptively in the hopes that this would make type inference easier. It does not make anything that much easier, because we don't, in fact, infer anything off of these Z nodes, and it actually undermines error-checking, because the resulting expressions under that Z node never set the type of any Z node, meaning that the type inference for such expressions is not done or lost.

This caused an issue where type specialization could allow complex guard tests through to the backend, where it would lead to a compile error.

To rectify this, we don't set the guard test type pre-emptively, and allow it to infer the type of the test expression through standard type inference. We already have a pass at the end of the parser pipeline to detect when guards have a non-array valued test expression, and we set G0 nodes to Z¯2 nodes when that happens.