Change the way we handle indexed assignment
This patch changes the way that we handle indexed assignment to align with the input that we get from the parser and to prepare for future extension to enable selective assignment.

The previous design assumed that we would have an indexed assignment of the from E4(N, I, F, V) where N was the name of the array being assigned, and I was the indexing array of type E6, and an assignment function F of the form P(←) or O(F, P(←)). These assignment functions would assume the presence of an E6 value as their left argument and the assignment target in their z position.

We are changing this so that we assume that any arbitrary expression will be in the indexing position I. We then choose the "indexing function" for the given primitive in that position. The assumption is that the chosen indexing function will provide the appropriate index set to the assignment function to work. In this case, the only indexing function we support is indexed assignment, which just verifies that the shapes are valid, and then returns the indices of the E6 expression. This means that the code generator works in exactly the same way as before, but we do a little bit more work in the APL side of the runtime, and we reduce the kernel size a little bit.

This sets the stage for more indexing functions to work with more assignment functions.

Because this analysis requires additional context, the easiest way to do this is to do the conversion from P → V earlier in the compiler before we have flattened everything out.