Add some code to manage the tagging of the output of the test
We can put the results of the test case at the bottom of the file in comments, but we have to do this in two stages. After we get the cases, we need to do a tagged output of the results, followed immediately by an annotation, which will give us the case code in annotated form.