Initial implementation of namespace resolution
This brings a basic implementation for namespace resolution into
the Resolve Names pass. It presently does not update type inference
to handle the results of a successful name resolution.

We primarily update the Resolve Names pass. For all the rest of the
code, we do not change the invariants of any of the AST except for the
lx field. The lx field changes to the following values after we complete
the Resolve Names pass:

  ¯4  Formals
  ¯3  Primitives
  ¯2  Free variables
  ¯1  Local variables
   0+ The node id of a variable's scope

In the case of definition bindings (H(V) nodes and T nodes), the lx of a
variable indicates that the node is bound to a namespace, and that the
variable itself serves as a valid left hand argument for namespace reference.
We expect all definition variables which are not namespace bound to have an
lx value of ¯1.

In the case of reference variables, the lx field indicates the scope of the
variable, which is local, free, or pointing to a specific namespace reference.

Resolutions of names occurs in exactly the same way that we have done
name resolution, except that we now check to see whether a variable is a
"scopee", meaning that it appears to the right of a dot. For such variables,
we require that the left hand side (the "scoper") be fully resolved before
we will attempt to resolve the scopee.

At the moment, we only consider scopees where the left hand side is a variable.
When the scoper variable is fully resolved and its definition site points to a
namespace (indicated with a non-negative lx field), then we use that namespace
as the scopee's scope and continue resolution.

We previously used a variable's referent (r) for its scope. Now, we define
a variable's "lookup referent" (lr) as being either r (lexical) or the value
lf a non-negative lx (namespace reference).