Don't specialize T nodes for bindings
It turns out that the actual semantics of Trad-fns and namespaces is that the name bound by the namespace or the trad-fn is not actually shadowed or bound locally in the body of either the namespace or the trad-fn. This means that when we are doing resolution, these are technically free variables and should be treated as such.