Fix handling of dynamic call sites
Because of changes to our handling of vb and how we deal with the lookup of trad-fns, we need to adjust our handling of dynamic closure call site recognition. To do this, we make a slight extension to the AST, to say that we will populate the vb of T nodes with the definition site variable where they are bound. We can then use this information to find such bindings and locate the definition closures that we need.