Refine the free variable propagation to be lexical/dynamic aware
Prior to this, we were just propagating all free variables through all the dynamic closures. This works okay, but is over-eager when it comes to lexical variables, and it fails to distinguish between lexical and dynamic variables.

This commit changes this. Now we only propagate dynamic variables to the dynamic closures, and we keep a separate variable in each closure for lexical and dynamic variables. Local nodes (H-bound) are still treated as the same thing (you cannot have a lexical and dynamically scoped variable resolve to different local bindings).

This necessitated bringing back the ¯3 value for lx. This means that we now use ¯2 for lexical variables, and ¯3 for dynamic variables. ¯1 is still used for locally bound variables that are not free.

This model we use here is that lexical functions resolve their free references at definition time, whereas dynamic bindings resolve their free references at call time using the call stack. This means, for instance, that free variables in a dfn defined inside of a trad-fns will reference whatever the dynamic binding is of the trad-fn at the time of execution, but calling a dfn from inside of a trad-fn will not change the bindings of its free variables. Likewise, calling a trad-fn that has free variables inside of a dfn will still resolve those variables dynamically, meaning that we need to propagate dynamically scoped variable out through the call stack of even a dfn.

Dyalog's interpreter currently gets this wrong, as it will switch whether a binding is dynamic or lexical on the basis of the next frame that it searchs, be it dynamic or lexical.

This implementation still has a fundamental limitation in that it may over-eagerly consider a variable free when an operator is called with a function containing dynamically free variables as an operator, but it shadows those variables locally in the body of the operator. In such cases, and in cases where we copy the function with something like f←fn or the like, we do not "push" the dynamic references to the actual call site, but only to the last known reference site.

Fixing this behavior will require unifying and tracing this behavior more thoroughly, but it is not considered a high priority at the moment. The only place where this could cause issues is if you wanted to check whether a function is ever passed with variables that have a different nameclass in different contexts. We might have a false positive in cases like the above, but this is yet to be seen in a meaningful way.