Some initial sketching of work on dynamic namespace resolution
The main limitations I see here are in how we detect the namespace expressions and the race condition between when a name is resolved and when we recognize that a declaration is "namespace assigned". If the namespace assignment occurs later than the resolution of some other variable, then that other variable may not trigger the right handling of scoped variables.